% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdata.R
\name{color_rating}
\alias{color_rating}
\title{Generate colors for TP ratings}
\usage{
color_rating()
}
\value{
Color palette function
}
\description{
Function to generate a leaflet palette function for coloring rating variable in leaflet maps
}
