% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdata.R
\name{filter_labels}
\alias{filter_labels}
\title{Filter troubled properties and create labels}
\usage{
filter_labels(mctroubledprops, filter = NULL)
}
\arguments{
\item{mctroubledprops}{Montgomery County troubled properties data, read and processed by read_fix_mc}

\item{filter}{Rating to which data is being filtered; options include "Troubled", "Compliant", "At-Risk", or "Tbd." Default NULL, for no filter}
}
\value{
A list of 1) the filtered dataset 2) the text supplied to the label argument of a leaflet map 3) the text supplied to the popup argument of a leaflet map
}
\description{
Optionally filters Montgomery County troubled properties dataset to a rating, and creates labels for circle-marker layers generated by markers_severity() or markers_severity_num()
}
