% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdata.R
\name{markers_severity}
\alias{markers_severity}
\title{Add Circle Markers of Building Rating}
\usage{
markers_severity(basemap, mctroubledprops, filter = NULL, sizeadjust = F)
}
\arguments{
\item{basemap}{Basemap adding layer to}

\item{mctroubledprops}{Dataframe of Montgomery County troubled properties}

\item{filter}{Filter to apply to mctroubledprops dataframe analyzing data/adding circle markers; default NULL}

\item{sizeadjust}{Whether to scale the size of the circle markers to the number of units in the property; default F}
}
\value{
Leaflet map with Circle-markers added
}
\description{
Adds circlemarkers of Montgomery County Troubled property building ratings to leaflet map
}
