% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{read_fix_mc}
\alias{read_fix_mc}
\title{Read and process MC Troubled Properties Data}
\usage{
read_fix_mc(apptoken)
}
\arguments{
\item{apptoken}{App token from Montgomery County's open-data developer portal to download troubled property data}
}
\value{
Latest-version of Montgomery-County troubled properties dataset, with broken addresses fixed
}
\description{
Read and process data from Montgomery County's Open data portal on inspections of troubled properties, fixing some addresses that are improperly geocoded. Data from here: https://data.montgomerycountymd.gov/Consumer-Housing/Troubled-Properties-Analysis/bw2r-araf
}
